+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [2.0.0-beta.3]
 ✨ New ✨
+* Added support for `dateFormat` to `MacosDatePicker`.
 * Added support for `startWeekOnMonday` to `MacosDatePicker`.
 
 🛠️ Fixed 🛠️